Separate source/header files, bug fixes…
- Non-templated, non-trivial or non-performance critical code is all in a separate source file.
- Replaced ELPP_STACKTRACE_ON_CRASH (which wasn’t used) with ELPP_FEATURE_CRASH_LOG.
- Small fixes for compiler warnings.
