+++ samples/MinGW/prog.cpp
@@ -1,5 +1,5 @@
 #define ELPP_STL_LOGGING
-// #define ELPP_STACKTRACE_ON_CRASH -- Stack trace not available for MinGW GCC
+// #define ELPP_FEATURE_CRASH_LOG -- Stack trace not available for MinGW GCC
 #define ELPP_PERFORMANCE_MICROSECONDS
 #define ELPP_LOG_STD_ARRAY
 #define ELPP_LOG_UNORDERED_MAP
