+++ samples/Qt/file-splitter-joiner/file-splitter-and-joiner.pro
@@ -15,7 +15,7 @@ DEFINES += ELPP_QT_LOGGING    \
           ELPP_FEATURE_ALL \
           ELPP_STL_LOGGING   \
           ELPP_STRICT_SIZE_CHECK \
-          ELPP_STACKTRACE_ON_CRASH \
+          ELPP_FEATURE_CRASH_LOG \
           ELPP_THREAD_SAFE
 
 COMPILER = g++
