+++ samples/STL/shared-static-libs/compile_static.sh
@@ -2,7 +2,7 @@ rm -rf libmyLib.so lib/libmyLib.so lib/mylib.o lib/myLib.a myLib.a myapp logs ##
 
 compiler=g++
 standard=c++0x ## If this does not work try c++11 (depends on your compiler)
-macros="-DELPP_THREAD_SAFE -DELPP_STACKTRACE_ON_CRASH"  ## Macros for library
+macros="-DELPP_THREAD_SAFE -DELPP_FEATURE_CRASH_LOG"  ## Macros for library
 
 cd lib/
 $compiler --std=$standard -pipe -fPIC -g -O0 $macros -Iinclude  -c -o mylib.o mylib.cpp
