+++ samples/STL/shared-storage/compile_shared.sh
@@ -2,11 +2,11 @@ rm -rf libmyLib.so lib/libmyLib.so lib/mylib.o lib/myLib.a myLib.a myapp logs ##
 
 compiler=icpc
 standard=c++0x ## If this does not work try c++11 (depends on your compiler)
-macros="-DELPP_THREAD_SAFE -DELPP_STACKTRACE_ON_CRASH"  ## Macros for library
+macros="-DELPP_THREAD_SAFE -DELPP_FEATURE_CRASH_LOG"  ## Macros for library
 
 cd lib/
 echo "$compiler --std=$standard -pipe -fPIC -g -O0  $macros  -Iinclude  -c mylib.cpp -o mylib.o"
-$compiler --std=$standard -pipe -fPIC -g -O0  $macros  -Iinclude  -c mylib.cpp -o mylib.o 
+$compiler --std=$standard -pipe -fPIC -g -O0  $macros  -Iinclude  -c mylib.cpp -o mylib.o
 echo "$compiler -fPIC -g  -shared -o libmyLib.so mylib.o"
 $compiler -fPIC -g  -shared -o libmyLib.so mylib.o
 cp libmyLib.so ..
