+++ test/easylogging++.h
@@ -7,7 +7,7 @@
 #define ELPP_STL_LOGGING
 #define ELPP_FORCE_ENV_VAR_FROM_BASH
 #define ELPP_ENABLE_THREADING
-#define ELPP_STACKTRACE_ON_CRASH
+#define ELPP_FEATURE_CRASH_LOG
 #define ELPP_SYSLOG
 #include "../src/easylogging++.h"
 #endif // EASYLOGGING_FOR_SAMPLES_H
